Fix varible name in eslint errors
