--- containers/LandingPage/LinerTwoImg/LinerTwoImg.module.scss
@@ -1,85 +1,0 @@
-.section {
-  display: -webkit-inline-box;
-  display: -ms-inline-flexbox;
-  display: inline-flex;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-  width: 100%;
-  margin: 0;
-  margin-top: 10px;
-  margin-bottom: 10px;
-  .one {
-    margin: 2.5px 5px 2.5px 2.5px;
-  }
-  .two {
-    margin: 2.5px 2.5px 2.5px 5px;
-  }
-  a {
-    display: block;
-    -webkit-box-flex: 1;
-    -ms-flex-positive: 1;
-    flex-grow: 1;
-    height: 100%;
-    // min-width: 24%;
-    border-radius: 9px;
-    overflow: hidden;
-    img {
-      width: 100%;
-      border-radius: 16px;
-    }
-  }
-}
-
-.section_desktop {
-  max-width: 1366px;
-  margin-right: auto;
-  margin-left: auto;
-  padding-right: 15px;
-  padding-left: 15px;
-  width: 100%;
-  &_hol {
-    margin-top: 17px;
-    .inner {
-      display: -ms-flexbox;
-      display: -webkit-box;
-      display: flex;
-      -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-      margin-right: 0;
-      margin-left: 0;
-    }
-    .container {
-      display: -webkit-inline-box;
-      display: -ms-inline-flexbox;
-      display: inline-flex;
-      -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-      justify-content: space-between;
-      width: 100%;
-      margin: 0;
-      a {
-        margin-left: 20px;
-        display: block;
-        -webkit-box-flex: 1;
-        -ms-flex-positive: 1;
-        flex-grow: 1;
-        margin-top: 20px;
-        div {
-          position: relative;
-          width: 100%;
-          float: left;
-          height: 100%;
-          img {
-            width: 100%;
-            height: 100%;
-            -webkit-box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.1);
-            box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.1);
-            cursor: pointer;
-            border-radius: 16px;
-          }
-        }
-      }
-    }
-  }
-}
